refactor(#73): extract shared SkillCopyButton
ExplorerSkillCard duplicated TacitKnowledgeView's copy button, and the two
copies diverged (the explorer's silently swallowed clipboard failures while
the tacit view downloaded a file). Extract one SkillCopyButton with the
clipboard→download fallback and a timer cleanup, used by both views.
